refactor(server): extract the last thin leaves → geo.js / ghost.js / promo.js
Pure move, zero behavior change. Clears the remaining standalone helpers
out of server.js, each into its own small module:

- src/server/geo.js     — normalizeGeoData (GEO Strategist data reshape;
                           pure transform)
- src/server/ghost.js   — buildGhostJWT (Ghost Admin HS256 JWT mint;
                           imports createHmac from crypto)
- src/server/promo.js   — PROMO_CODES (static comp-code map)

Why: finishes the clean-leaf phase of the decomposition before the
route-group surgery. Each was a single-use standalone with no coupling to
the rest of the monolith.

What: three new modules, each exporting its one symbol (moved verbatim);
server.js imports them; inline defs -> breadcrumbs.

Test plan:
- node --check server.js -> OK
- npm run lint (no-undef gate) -> clean (all three resolve; no orphaned refs)
- route-inventory guard -> PASS (snapshot unchanged)
- vitest -> 86/86 pass (+8 new: normalizeGeoData priority bucketing /
  platform mapping / brief fallbacks; buildGhostJWT segment+header+payload
  shape; PROMO_CODES lookup)

Rollback: revert; pure copies, restore inline defs exactly.
