+++ src/server/geo.js
@@ -1,0 +1,55 @@
+// GEO data normalization, extracted from server.js during the decomposition.
+// normalizeGeoData reshapes the GEO Strategist's raw brief / topical map / GEO
+// opportunities / entity schema into the normalized structure the API returns.
+// Pure transform — no DB/network.
+
+export function normalizeGeoData(briefData, topicalMap, geoOpportunities, entitySchema, profile) {
+  const gaps = (topicalMap && topicalMap.gapsByCluster) || [];
+  if (gaps.length > 0) console.log('[GEO] normalizer gaps[0] RAW:', JSON.stringify(gaps[0]));
+  const topicalAuthorityMap = gaps.map(g => {
+    const score = g.geoCitationScore || g.citationProbability || g.score || g.geoScore || g.probability || 0;
+    return {
+      topic: g.topic || g.cluster || g.name || g.title || 'Unknown',
+      coverage: g.rationale || g.description || g.reason || g.owner || g.gap || '',
+      citationProbability: score,
+      priority: score >= 70 ? 'high' : score >= 40 ? 'medium' : 'low'
+    };
+  });
+
+  const topicMap = {};
+  (geoOpportunities || []).forEach(o => {
+    const t = o.topic || 'Unknown';
+    if (!topicMap[t]) topicMap[t] = { topic: t, chatgpt: 0, perplexity: 0, aiOverviews: 0, gemini: 0, quickWin: o.quickWin || false };
+    const p = (o.platform || '').toLowerCase().replace(/\s/g, '');
+    if (p.includes('chatgpt') || p.includes('openai')) topicMap[t].chatgpt = o.score || 0;
+    else if (p.includes('perplexity')) topicMap[t].perplexity = o.score || 0;
+    else if (p.includes('overview') || p.includes('google')) topicMap[t].aiOverviews = o.score || 0;
+    else if (p.includes('gemini')) topicMap[t].gemini = o.score || 0;
+    if (o.quickWin) topicMap[t].quickWin = true;
+  });
+  const geoOpportunitiesNorm = Object.values(topicMap);
+
+  const entitySchemaMap = (entitySchema || []).map(e => ({
+    entity: e.entity || '',
+    schemaType: Array.isArray(e.schemaTypes) ? e.schemaTypes[0] : (e.schemaType || 'Article'),
+    competitorCited: e.competitorCiting || e.competitorCited || false,
+    recommendation: e.rationale || e.recommendation || ''
+  }));
+
+  const h2sRaw = briefData.h2s || [];
+  const geoBrief = {
+    title: briefData.titleTag || briefData.title || briefData.targetTopic || (profile && profile.brand_name) || '',
+    h1: briefData.h1 || briefData.targetTopic || '',
+    h2s: h2sRaw.map(h => typeof h === 'string' ? h : h.heading || h.h2 || ''),
+    faqItems: (briefData.faqStructure || briefData.faqItems || []).map(f => ({
+      q: f.question || f.q || '',
+      a: f.answerDirection || f.answer || f.a || ''
+    })),
+    geoAnchors: briefData.geoAnchors || [],
+    estimatedCitationLift: briefData.geoScorecard
+      ? `+${Math.round((briefData.geoScorecard.currentReadiness || 0) * 0.4)}% in 90 days`
+      : '+15–30% in 90 days'
+  };
+
+  return { topicalAuthorityMap, geoOpportunities: geoOpportunitiesNorm, entitySchemaMap, geoBrief };
+}
